【PR】: sync develop to master 0519

Created-by: zhujingjing
Commit-by: wuzheng-hw;lidaoming1;coconut_tree;du-hua1024;xingzhixiong;maohaodi;tangqunzhang;xiangrenzhi;AlexRen998916;yelongjian;jikai-tyler;A-Liuhao;kobemini;Renzhi Xiang;叶龙剑;Rentianqi
Merged-by: zhujingjing
Description: 

# Pull Request

## 描述
请清晰准确地描述本次 Pull Request 的意图和变更内容。
1、修复 StatusFactory 单例及 DSO 跨库访问问题
2、清理代码中的编译 warning
3、为 graph 模块新增 stub API 支持
4、修复 sleep 接口相关 bug
5、解耦 llm_datadist v1 与 v2 接口
6、修复文档中的拼写错误和格式问题（关联 issue #379 #380）
7、修复 codecheck 告警问题
8、将 device/stream 相关 runtime 接口重构为 aclrt 接口
9、清理 runtime 相关代码的 const 正确性
10、修复 tiny 模式用例执行失败及 autofuse 集成问题
11、ACL 模型层优化与重构，拆分 libacl_mdl_impl_om2.so
12、修复 Python C API 的 const 声明不一致问题（G.FUN.02-CPP 规范）
13、RT2场景下调用用户实现的infershape函数
14、memcpy_addr_async om2适配codegen修改
## 变更类型
请选择本次引入的变更类型：
<!--  [x] 表示选中 -->
- [x] 🐛 Bug 修复
- [ ] ✨ 新功能
- [x] 💄 代码风格更新（格式化，局部变量）
- [ ] ♻️ 重构（既不修复错误也不增加功能的代码变动）
- [ ] 📦 构建过程或辅助工具的变动
- [x] 📝 文档内容更新

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在当前页面的右侧'关联Issue'部分添加相应Issue链接，并勾选'合并后关闭已关联的 Issue'选项。 -->


## 如何测试
描述测试此变更的步骤和前提条件：
1.
2.

## 核对清单
<!--  [x] 表示选中 -->
- [x] 我的代码遵循了项目的代码风格
- [x] 我已对代码进行了自测
- [x] 我已更新了相关的文档
- [x] 我在标题中使用了合适的类型标签（如：`feat:`, `fix:`）
- [x] 我已经详细阅读了贡献指南（CONTRIBUTING.md），并遵守了其中的所有规定，包括但不限于commit message的格式、无效commit的合并等

## 其他信息
在此添加任何其他关于本次 PR 的说明。




See merge request: cann/ge!2986

+++ dflow/pydflow/wrapper/dflow_wrapper.cpp
@@ -439,7 +439,7 @@ void BindFlowInfo(py::module &m) {
 }
 
 ge::Tensor CreateTensorFromNumpyArray(const py::array &np_array) {
-  auto flags = static_cast<unsigned int>(np_array.flags());
+  auto flags = np_array.flags();
   if ((flags & pybind11::detail::npy_api::NPY_ARRAY_C_CONTIGUOUS_) == 0) {
     throw std::runtime_error("Numpy array is not C Contiguous");
   }
